first commit of the iOS version of vectorboolean. need to write some tests to confirm functionality didn't change in the refactor. bonus points later to make this ARC

+++ VectorBoolean/FBDebug.m
@@ -15,7 +15,7 @@ NSString *FBArrayDescription(NSArray *array)
     for (NSUInteger i = 0; i < [array count]; i++) {
         if ( i == 0 )
             [description appendString:@"\n"];
-        [description appendFormat:@"\t%lu\t=\t%@\n", i, [array objectAtIndex:i]];
+        [description appendFormat:@"\t%lu\t=\t%@\n", (unsigned long) i, [array objectAtIndex:i]];
     }
     [description appendString:@"]"];
     return description;
